+++ gen/template/unistrip.cpp
@@ -12,7 +12,6 @@
 
 #include "unistrip.h"
 
-namespace ufal {
 namespace unilib {
 
 const uint8_t unistrip::combining_mark_index[unistrip::CHARS >> 8] = $COMBINING_MARK_INDICES;
@@ -26,4 +25,3 @@ const uint16_t unistrip::stripped_block[][256] = $STRIPPED_BLOCKS;
 const char32_t unistrip::stripped_data[] = $STRIPPED_DATA;
 
 } // namespace unilib
-} // namespace ufal